RA-35 refactor 1
