+++ app/src/main/java/com/example/myrecipeapp/ui/recipes/list_recipes/RecipeListViewModel.kt
@@ -7,7 +7,7 @@ import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.viewModelScope
-import com.example.myrecipeapp.RecipesRepository
+import com.example.myrecipeapp.data.RecipesRepository
 import com.example.myrecipeapp.model.Category
 import com.example.myrecipeapp.model.Recipe
 import kotlinx.coroutines.launch
